project/requirements.txt
fastapi
uvicorn
gunicorn
SQLAlchemy
psycopg2-binary
pydantic
python-dotenv
passlib[bcrypt]
python-jose
boto3
requests